+++ src/clojure/nrepl/logging.clj
@@ -3,14 +3,20 @@
   {:added  "0.6.0"})
 
 ;; TODO: Use clojure.tools.logging if available
-(def default-output *out*)
+(def stdout *out*)
+(def stderr *err*)
 
 (defn info
   [& args]
-  (binding [*out* default-output]
+  (binding [*out* stdout]
     (apply println "INFO: " args)))
 
 (defn warn
   [& args]
-  (binding [*out* default-output]
+  (binding [*out* stderr]
     (apply println "WARN: " args)))
+
+(defn error
+  [& args]
+  (binding [*out* stderr]
+    (apply println "ERROR: " args)))