+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### master (unreleased)
 
+#### New features
+
+* [#85](https://github.com/nrepl/nrepl/issues/85): Implement message logging functionality.
+
 ### 0.5.2 (2018-12-10)
 
 #### Bugs fixed
@@ -22,7 +26,6 @@
   function in `pr-values`, enabling pretty-printed REPL results.
 * [#66](https://github.com/nrepl/nrepl/pull/66): Add support for a global and local configuration file.
 * [#63](https://github.com/nrepl/nrepl/issues/63): Make it possible to specify the transport via the command-line client (`--transport/-t`).
-* [#85](https://github.com/nrepl/nrepl/issues/85): Implement message logging functionality.
 
 #### Bugs fixed